+++ R/rlistings.R
@@ -140,8 +140,10 @@ as_listing <- function(df,
                        main_footer = NULL,
                        prov_footer = NULL) {
   if (length(non_disp_cols) > 0 && length(intersect(key_cols, non_disp_cols)) > 0) {
-    stop("Key column also listed in non_disp_cols. All key columns are by",
-         " definition display columns.")
+    stop(
+      "Key column also listed in non_disp_cols. All key columns are by",
+      " definition display columns."
+    )
   }
   if (!is.null(disp_cols) && !is.null(non_disp_cols)) {
     stop("Got non-null values for both disp_cols and non_disp_cols. This is not supported.")
@@ -153,13 +155,16 @@ as_listing <- function(df,
     cols <- disp_cols
   }
   if (!all(sapply(default_formatting, is, class2 = "fmt_config"))) {
-    stop("All format configurations supplied in `default_formatting`",
-         " must be of type `fmt_config`.")
+    stop(
+      "All format configurations supplied in `default_formatting`",
+      " must be of type `fmt_config`."
+    )
   }
-  if (!(is.null(col_formatting) ||
-        all(sapply(col_formatting, is, class2 = "fmt_config")))) {
-    stop("All format configurations supplied in `col_formatting`",
-         " must be of type `fmt_config`.")
+  if (!(is.null(col_formatting) || all(sapply(col_formatting, is, class2 = "fmt_config")))) {
+    stop(
+      "All format configurations supplied in `col_formatting`",
+      " must be of type `fmt_config`."
+    )
   }
 
   df <- as_tibble(df)
@@ -182,20 +187,27 @@ as_listing <- function(df,
   ## key cols must be leftmost cols
   cols <- c(key_cols, setdiff(cols, key_cols))
 
+  row_all_na <- apply(df[cols], 1, function(x) all(is.na(x)))
+  if (any(row_all_na)) {
+    message("rows that only contain NA values have been trimmed")
+    df <- df[!row_all_na, ]
+  }
+
   # set col format configs
   df[cols] <- lapply(cols, function(col) {
     col_class <- tail(class(df[[col]]), 1)
-    col_fmt_class <- if (!col_class %in% names(default_formatting) &&
-                         is.numeric(df[[col]])) "numeric" else col_class
+    col_fmt_class <- if (!col_class %in% names(default_formatting) && is.numeric(df[[col]])) "numeric" else col_class
     col_fmt <- if (col %in% names(col_formatting)) {
       col_formatting[[col]]
     } else if (col_fmt_class %in% names(default_formatting)) {
       default_formatting[[col_fmt_class]]
     } else {
       if (!"all" %in% names(default_formatting)) {
-        stop("Format configurations must be supplied for all listing columns. ",
-             "To cover all remaining columns please add an 'all' configuration",
-             " to `default_formatting`.")
+        stop(
+          "Format configurations must be supplied for all listing columns. ",
+          "To cover all remaining columns please add an 'all' configuration",
+          " to `default_formatting`."
+        )
       }
       default_formatting[["all"]]
     }
@@ -269,7 +281,6 @@ setMethod(
     atts <- attributes(obj)
     atts$names <- cols
     attributes(listing) <- atts
-
     keycols <- get_keycols(listing)
 
 
@@ -285,6 +296,7 @@ setMethod(
     for (i in seq_along(keycols)) {
       kcol <- keycols[i]
       kcolvec <- listing[[kcol]]
+      kcolvec <- vapply(kcolvec, format_value, "", format = obj_format(kcolvec), na_str = obj_na_str(kcolvec))
       curkey <- paste0(curkey, kcolvec)
       disp <- c(TRUE, tail(curkey, -1) != head(curkey, -1))
       bodymat[disp, kcol] <- kcolvec[disp]
@@ -313,6 +325,7 @@ setMethod(
         byrow = TRUE
       )
     )
+
     MatrixPrintForm(
       strings = fullmat,
       spans = matrix(1,
@@ -397,7 +410,7 @@ add_listing_col <- function(df,
                             na_str = "NA",
                             align = "left") {
   if (!is.null(fun)) {
-    vec <- fun(df)
+    vec <- with_label(fun(df), name)
   } else if (name %in% names(df)) {
     vec <- df[[name]]
   } else {
@@ -408,7 +421,6 @@ add_listing_col <- function(df,
   }
 
   if (!is.null(format)) {
-    vec <- df[[name]]
     obj_format(vec) <- format
   }